refactored code after code review. Updates to server.js. Added controllers
