+++ controllers/image.js
@@ -1,0 +1,11 @@
+export const handelEntries = (req, res, db) => {
+	const { id } = req.body;
+	db('users')
+		.where('id', '=', id)
+		.increment({ entries: 1 })
+		.returning('entries')
+		.then((entries) => {
+			res.json(entries[0].entries);
+		})
+		.catch((err) => res.status(400).json('Unable to get entries'));
+}
